### 2025-04-08 8651c2d
updated missing DOI at paper.bib
