+++ paper/paper.bib
@@ -69,7 +69,8 @@
   author = {Dario Azzimonti and Nicolò Rubattu and Lorenzo Zambon and Giorgio Corani},
   year = {2024},
   note = {R package version 0.3.2},
-  url = {https://idsia.github.io/bayesRecon/}
+  url = {https://idsia.github.io/bayesRecon/},
+  doi = {10.32614/CRAN.package.bayesRecon}
 }
 
 @manual{hts_Hyndman,
@@ -77,7 +78,8 @@
   author = {Rob Hyndman and Alan Lee and Earo Wang and Shanika Wickramasuriya},
   year = {2024},
   note = {R package version 6.0.3},
-  url = {https://CRAN.R-project.org/package=hts}
+  url = {https://CRAN.R-project.org/package=hts},
+  doi = {10.1016/j.csda.2011.03.006}
 }
 
 @manual{thief_HyndmanKourentzes2018,
@@ -85,7 +87,8 @@
   author = {Rob J Hyndman and Nikolaos Kourentzes},
   year = {2018},
   note = {R package version 0.3},
-  url = {http://pkg.robjhyndman.com/thief}
+  url = {http://pkg.robjhyndman.com/thief},
+  doi = {10.32614/CRAN.package.thief}
 }
 
 @article{olivares2022hierarchicalforecast,
@@ -93,7 +96,8 @@
   author = {Olivares, Kin G. and Garza, Federico and Luo, David and Challú, Cristian and Mergenthaler, Max and Ben Taieb, Souhaib and Wickramasuriya, Shanika L. and Dubrawski, Artur},
   journal = {arXiv preprint arXiv:2207.03517},
   year = {2022},
-  url = {https://arxiv.org/abs/2207.03517}
+  url = {https://arxiv.org/abs/2207.03517},
+  doi = {10.48550/arXiv.2207.03517}
 }
 
 @article{gluonTS_Alexandrov_etal2020,
@@ -104,7 +108,8 @@
   number = {116},
   pages = {1--6},
   year = {2020},
-  url = {http://jmlr.org/papers/v21/19-820.html}
+  url = {http://jmlr.org/papers/v21/19-820.html},
+  doi = {10.5555/3455716.3455832}
 }
 
 @Manual{smooth_pkg,
@@ -112,7 +117,8 @@
   author = {Ivan Svetunkov},
   year = {2024},
   note = {R package version 4.1.0},
-  url = {https://CRAN.R-project.org/package=smooth}
+  url = {https://CRAN.R-project.org/package=smooth},
+  doi = {10.1201/9781003452652}
 }
 
 @manual{FoReco,
@@ -129,8 +135,10 @@
   author = {Mitchell O'Hara-Wild and Rob Hyndman and Earo Wang},
   year = {2024},
   note = {R package version 0.4.0},
-  url = {https://fable.tidyverts.org}
+  url = {https://fable.tidyverts.org},
+  doi = {10.32614/CRAN.package.fable}
 }
+
 
 @article{ZHANG2024143,
   title = {Discrete forecast reconciliation},
