### 2025-04-08 6648627
updated missing DOI at paper.bib
