+++ paper/paper.bib
@@ -108,8 +108,7 @@
   number = {116},
   pages = {1--6},
   year = {2020},
-  url = {http://jmlr.org/papers/v21/19-820.html},
-  doi = {10.5555/3455716.3455832}
+  url = {https://dl.acm.org/doi/10.5555/3455716.3455832}
 }
 
 @Manual{smooth_pkg,
@@ -173,14 +172,15 @@
   doi = {10.1016/j.ijforecast.2023.10.003}
 }
 
-@article{taieb2021hierarchical,
+@article{taieb2020hierarchical,
   title = {Hierarchical probabilistic forecasting of electricity demand with smart meter data},
-  author = {Taieb, Souhaib Ben and Taylor, James W and Hyndman, Rob J},
+  author = {Taieb, Souhaib Ben and Taylor, James W},
   journal = {Journal of the American Statistical Association},
-  volume = {116},
-  number = {533},
-  pages = {27--43},
-  year = {2021}
+  volume = {115},
+  number = {531},
+  pages = {1223--1235},
+  year = {2020},
+  doi = {10.1080/01621459.2020.1736081}
 }
 
 @article{svetunkov2023iets,
