### 2025-04-08 a5443b3
updated missing DOI at paper.bib
