+++ paper/paper.bib
@@ -16,7 +16,8 @@
   volume = {262},
   number = {1},
   pages = {60--74},
-  year = {2017}
+  year = {2017},
+  doi = {10.1016/j.ejor.2017.02.046}
 }
 
 @article{Hyndman_Ahmed_Athanasopoulos_Shang_2011,
@@ -240,12 +241,13 @@
 }
 
 @inproceedings{zambon2024probabilistic,
-  title = {Probabilistic reconciliation of mixed-type hierarchical time series},
+  title = {Probabilistic Reconciliation of Mixed-Type Hierarchical Time Series},
   author = {Zambon, Lorenzo and Azzimonti, Dario and Rubattu, Nicolò and Corani, Giorgio},
   booktitle = {Proceedings of the Fortieth Conference on Uncertainty in Artificial Intelligence},
   pages = {4078--4095},
   year = {2024},
   editor = {Kiyavash, Negar and Mooij, Joris M.},
   volume = {244},
-  series = {PMLR}
+  series = {Proceedings of Machine Learning Research},
+  url = {https://proceedings.mlr.press/v244/zambon24a.html}
 }
