### 2025-04-14 b07661e
Updated paper.md with a lower word count.
