+++ paper/paper.md
@@ -50,20 +50,21 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Forecast reconciliation is essential for ensuring coherence across hierarchical time series, where aggregate values must match the sum of their components. A common example is retail sales: item-level sales can be aggregated by product, category, or store [@MAKRIDAKIS20221325]. While the observed data are coherent by construction, independently generated forecasts (base forecasts) typically violate aggregation constraints, resulting in incoherent predictions.
+Forecast reconciliation ensures coherence across hierarchical time series, where aggregate forecasts must equal the sum of their components. In practice, base forecasts—generated independently for each series—often violate these constraints.
 
-Early work addressed this issue for point forecasts using projection-based methods such as OLS and MinT [@Hyndman_Ahmed_Athanasopoulos_Shang_2011; @Wickramasuriya_Athanasopoulos_Hyndman_2019]. More recently, probabilistic reconciliation methods have been proposed to reconcile entire forecast distributions [@Jeon_Panagiotelis_Petropoulos_2019; @taieb2021hierarchical; @Panagiotelis_Gamakumara_Athanasopoulos_Hyndman_2023; @Girolimetto_Athanasopoulos_DiFonzo_Hyndman_2024], offering a richer and more informative forecasting framework.
+Initial approaches addressed point forecasts using projection methods like OLS and MinT [@Hyndman_Ahmed_Athanasopoulos_Shang_2011; @Wickramasuriya_Athanasopoulos_Hyndman_2019]. More recently, probabilistic reconciliation methods have been introduced, providing richer uncertainty quantification [@Jeon_Panagiotelis_Petropoulos_2019; @Panagiotelis_Gamakumara_Athanasopoulos_Hyndman_2023].
 
-Despite this progress, most existing software tools for reconciliation have significant limitations (**Table 1**). Many only support Gaussian or continuous distributions, lack support for discrete or mixed-type forecasts, or are implemented only in R. Others, such as ProbReco [@Panagiotelis_Gamakumara_Athanasopoulos_Hyndman_2023] and DiscreteRecon [@ZHANG2024143], are no longer actively maintained. Notably, no existing Python library offers a unified interface for probabilistic reconciliation across distribution types.
+However, most existing tools are limited to Gaussian or continuous inputs, lack support for discrete or mixed-type forecasts, or are implemented only in R. Some, like ProbReco and DiscreteRecon, are no longer actively maintained.
 
-`bayesReconPy` fills this gap. It is a Python package that implements probabilistic reconciliation via both conditioning and projection-based methods, including OLS and MinT. It supports a broad range of forecast types:
+`bayesReconPy` addresses these gaps. It provides a unified Python interface for probabilistic reconciliation using both conditioning and projection-based methods. It supports:
 
-- Gaussian forecasts [@Corani_Azzimonti_Augusto_Zaffalon_2021]
-- Continuous non-Gaussian forecasts [@Zambon_Azzimonti_Corani_2024]
-- Discrete forecasts [@CoraniAzzimontiRubattu2024; @Zambon_Azzimonti_Corani_2024]
+- Gaussian forecasts [@Corani_Azzimonti_Augusto_Zaffalon_2021]  
+- Continuous non-Gaussian forecasts [@Zambon_Azzimonti_Corani_2024]  
+- Discrete forecasts [@CoraniAzzimontiRubattu2024]  
 - Mixed discrete-continuous hierarchies [@zambon2024probabilistic]
 
[…]
