### 2025-04-16 #8022 @goldingn
Hey @mitchelloharawild or @robjhyndman this is in python not R, and you may both be too busy, but any chance either of you would be interested in reviewing the software and short paper for JOSS?
