### 2025-04-16 #8022 @goldingn
Hi @pitmonticone would you be interested in reviewing this JOSS paper and Python package on Bayesian forecast reconciliation?
