### 2025-04-16 #8022 @robjhyndman
> Hey [@mitchelloharawild](https://github.com/mitchelloharawild) or [@robjhyndman](https://github.com/robjhyndman) this is in python not R, and you may both be too busy, but any chance either of you would be interested in reviewing the software and short paper for JOSS?

Sorry, my knowledge of Python is not really sufficient for reviewing software. You could try @cbergmeir or @kdgutier
