### 2025-04-17 #8022 @goldingn
Thanks Rob! @cbergmeir or @kdgutier would you be interested in reviewing this JOSS paper and Python package on Bayesian forecast reconciliation?
