### 2025-05-07 #8022 @goldingn
Hi @pitmonticone, @wsavran, @kdgutier, I'm just following up - would you be able to review this Python package submission on Bayesian forecast reconciliation for JOSS?
