### 2025-05-26 #8022 @earthai-tech
> Hi [@earthai-tech](https://github.com/earthai-tech), would you be interested in reviewing this python package on forecast reconciliation for JOSS?

Hi @goldingn,
Yes, I’d be happy to review it. Looking forward to contributing!
