### 2025-05-27 #8022 @robmoss
> Hi [@robmoss](https://github.com/robmoss), would you be interested in reviewing this python package on forecast reconciliation for JOSS?

Yes, this looks like an interesting package to review! I'll need a few weeks, I have a few other pressing commitments ATM. I hope that's okay.
