### 2025-06-03 #8336 @editorialbot
[REVIEW]: BayesReconPy: A Python package for forecast reconciliation
**Submitting author:** @Anubhabsupsi (<a href="http://orcid.org/0009-0001-2780-7555">Anubhab Biswas</a>)
**Repository:** https://github.com/supsi-dacd-isaac/BayesReconPy
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.3
