### 2025-06-03 432f754
Updated paper.md with a typo-fix.
