### 2025-06-17 supsi-dacd-isaac/BayesReconPy#1 @robmoss
JOSS paper: statement of need / state of the field
Hi there! I've begun [my review](https://github.com/openjournals/joss-reviews/issues/8336#issuecomment-2938200144) of your paper, and I've done a quick search to see what other reconciliation packages are available for Python. The following packages seem relevant, and most of them appear to be active:

- [hierarchicalforecast](https://github.com/Nixtla/hierarchicalforecast)
- [probabilistic-reconciliation](https://github.com/dirmeier/reconcile)
- [pyhts](https://github.com/AngelPone/pyhts)
- [scikit-hts](https://github.com/carlomazzaferro/scikit-hts)

In the **Statement of Need** section, could you please briefly compare `BayesReconPy` to these packages?
