### 2025-06-17 supsi-dacd-isaac/BayesReconPy#2 @robmoss
JOSS paper: reproducing the notebook results
As part of [my review](https://github.com/openjournals/joss-reviews/issues/8336#issuecomment-2938200144) I've run all of the examples included in the three notebooks. This involved installing `matplotlib`, `properscoring`, and `tabulate`, which are not dependencies of `BayesReconPy`. I think it's probably worth adding this to the installation instructions in the "Probabilistic Reconciliation via Conditioning with bayesreconpy" notebook.

I was able to reproduce all results **except** for the car part sales example in the "Probabilistic Reconciliation via Conditioning with bayesreconpy" notebook.

> We now compute the Mean Absolute Error (MAE) and the Continuous Ranked Probability Score (CRPS) for the bottom (i.e., monthly) time series. For computing CRPS, we use the package `properscoring`.

At this point the Python session was killed due to an out-of-memory exception (on my laptop with 16 GB RAM). I tried re-running it for only a single point in the forecast horizon and still experienced this issue. I have no concerns about the validity of the results, given that I was able to reproduce everything else in the notebooks, I'm just surprised at the extremely high RAM usage. Perhaps it's worth remarking in the notebook?
