### 2025-06-17 supsi-dacd-isaac/BayesReconPy#3 @robmoss
JOSS paper: online documentation
As part of [my review](https://github.com/openjournals/joss-reviews/issues/8336#issuecomment-2938200144), I've looked at the [online documentation](https://bayesreconpy.readthedocs.io/en/latest/). Installation instructions are provided and the core functionality is documented, including example usage. It looks good! With reference to the review criteria, the authors may wish to add:

- A short statement of need ("what problems the software is designed to solve and who the target audience is"). The documentation only includes a link to the [original R package](https://cran.r-project.org/web/packages/bayesRecon/index.html), which itself has only a very short blurb, and while the repository `README.md` includes an introduction, it relies on the reader already understanding what problems the software is designed to solve.

- Guidelines for third parties wishing to contribute to the software, report issues or problems with the software, and seek support. This might be as simple as adding a few sentences to the `README.md`, or adding a separate `CONTRIBUTING.md` file (here is [one such example](https://github.com/desResLab/LINFA/blob/master/CONTRIBUTING.md)).
