+++ requirements.txt
@@ -3,6 +3,9 @@ pandas>=2.2.2
 PuLP>=2.9.0
 scipy>=1.14.1
 KDEpy>=1.1.10
+matplotlib==3.9.2
+properscoring==0.1
+tabulate==0.9.0
 pydata-sphinx-theme
 sphinx-autodoc-typehints
 nbsphinx
