### 2025-06-27 supsi-dacd-isaac/BayesReconPy#1 @Anubhabsupsi
Hi @robmoss, thanks a lot for the detailed comments and for reviewing our submission. I was already familiar with the other three packages, but I wasn't aware of the [probabilistic-reconciliation](https://github.com/dirmeier/reconcile) package — really appreciate the pointer. I’ll update the statement of need to include them all soon, and update you.
