### 2025-06-27 supsi-dacd-isaac/BayesReconPy#2 @Anubhabsupsi
Hi @robmoss , thanks for pointing it out. My PC has a 32GB RAM and I didn't face any trouble, but now that I ran the code again and had a look at the RAM history it seems obvious that running it on a 16GB RAM will kill the session. The notebook however works well with google colab, and we are trying to fix this issue by re-writing the notebook. I'll update you again as soon as we have the fix.
