### 2025-06-27 supsi-dacd-isaac/BayesReconPy#2 @robmoss
Fair enough, thanks very much for investigating options! My surprise about RAM usable is mostly, or entirely, because it felt like a reasonably small-scale example.
