+++ CONTRIBUTING.md
@@ -1,0 +1,224 @@
+# Contributing to BayesReconPy 🧠
+
+First off, thanks for considering contributing! These guidelines will help you get started.
+
+---
+
+## 📚 Table of Contents
+
+- [🧰 Getting Started](#-getting-started)
+  - [Requirements](#requirements)
+  - [Code Style & Formatting](#code-style--formatting)
+- [🧪 Running Tests](#-running-tests)
+- [🚀 Pull Request Workflow](#-pull-request-workflow)
+  - [PR Guidelines](#pr-guidelines)
+- [🔧 Reporting Bugs & Requesting Features](#-reporting-bugs--requesting-features)
+- [📐 Code Guidelines](#-code-guidelines)
+- [📚 Documentation](#-documentation)
+- [✅ Review & Merge Process](#-review--merge-process)
+- [🏷️ Versioning & Releases](#️-versioning--releases)
+- [🤝 Code of Conduct](#-code-of-conduct)
+- [🙏 Thank You!](#-thank-you)
+
+---
+
+## 🧰 Getting Started
+
+### Requirements
+
+Make sure you have:
+
+- Python 3.8+ (preferably via `venv` or `conda`)
+- Core dependencies installed via:
+
+  ```bash
+  pip install -e .[dev]
+  ```
+
+## 🎨 Code Style & Formatting
+
+We enforce consistent and readable code using automatic tools.
+
+### 📦 Required Tools
+
+Make sure you have the development dependencies installed:
+
+```bash
+pip install -e .[dev]
+```
+
+Then install pre-commit hooks (run once):
+
+```bash
+pip install pre-commit
+pre-commit install
+```
+
+These tools ensure that code is formatted and linted before each commit.
+
+---
+
+### ✨ Formatting with `black`
+
+We use [`black`](https://black.readthedocs.io/) for code formatting:
+
+```bash
+black .
+```
+
+`black` will automatically reformat your code to follow a consistent style.
+
+---
+
+### 🧹 Linting with `flake8`
+
+We use [`flake8`](https://flake8.pycqa.org/) for static analysis and code linting:
+
+```bash
+flake8 bayesreconpy
+```
+
+This helps catch syntax errors, unused variables, and style violations.
+
+---
+
+### ⚙️ Recommended Configuration
+
+You can add the following configuration to your `pyproject.toml` or `.flake8` file to customize checks:
+
+#### `.flake8`
+
+```ini
+[flake8]
+max-line-length = 88
+extend-ignore = E203, W503
+```
+
+#### `pyproject.toml`
+
+```toml
+[tool.black]
+line-length = 88
+target-version = ['py38']
+```
+
+---
+
+### 💡 Tip
+
+All of this is automatically run if you commit via `git` and have `pre-commit` installed. To manually run all hooks:
+
+```bash
+pre-commit run --all-files
+```
+
+---
+
+Following these steps helps ensure code quality and consistency across contributions.
+
+---
+
+## 🧪 Running Tests
+
+Use `pytest` to run the test suite:
+
+```bash
+pytest
+```
+
+For test coverage reports:
+
+```bash
+pytest --cov=bayesreconpy
+```
+
+---
+
+## 🚀 Pull Request Workflow
+
+1. Fork the repository and clone your fork.
+2. Create a new branch:
+
+   ```bash
+   git checkout -b feature/your-feature-name
+   ```
+
+3. Implement your changes and add tests.
+4. Ensure all checks pass (`black`, `flake8`, and `pytest`).
+5. Commit your changes:
+
+   ```bash
+   git add .
+   git commit -m "feat: Add your concise feature summary"
+   ```
+
+6. Push your branch and open a Pull Request against the `main` branch.
+
+### PR Guidelines
+
[…]
