### 2025-07-07 5b3a7cf
Updated notebook 1 with a different crps function and an image for the online documentation
