+++ requirements.txt
@@ -4,7 +4,6 @@ PuLP>=2.9.0
 scipy>=1.14.1
 KDEpy>=1.1.10
 matplotlib==3.9.2
-properscoring==0.1
 tabulate==0.9.0
 pydata-sphinx-theme
 sphinx-autodoc-typehints
