### 2025-07-07 supsi-dacd-isaac/BayesReconPy#2 @Anubhabsupsi
Hi @robmoss , I have added a new function for CRPS and now it should be okay - the implementation is fast as well now. I have also updated the installation instructions as you suggested. Let me know if you still face an issue with the RAM.
