### 2025-07-07 7e2e3d6
updated the python version in CONTRIBUTING.md
