### 2025-07-07 supsi-dacd-isaac/BayesReconPy#3 @Anubhabsupsi
Hi @robmoss thanks for the valuable suggestions. I have added a brief description in the online documentation, for the first bullet point. For the second one, I have added a `CONTRIBUTING.md` file to the package as welll.
