### 2025-07-08 f217710
minor change - added pypi versions
