### 2025-07-08 e9dd3ff
new version 0.3.0 - change of API
