+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "bayesreconpy"
-version = "0.2.0"
+version = "0.3.0"
 description = "Bayesian reconciliation for hierarchical forecasting"
 readme = "README.md"
 requires-python = ">=3.10"
