+++ paper/paper.bib
@@ -139,6 +139,23 @@
   doi = {10.32614/CRAN.package.fable}
 }
 
+@misc{scikit-hts,
+  author       = {C. Jason Ross},
+  title        = {scikit-hts: Hierarchical Time Series Forecasting with Python},
+  year         = {2019},
+  publisher    = {GitHub},
+  journal      = {GitHub repository},
+  howpublished = {\url{https://github.com/cjayross/scikit-hts}},
+}
+
+@misc{dirmeier2025reconcile,
+  author       = {Simon Dirmeier},
+  title        = {reconcile: Hierarchical Time Series Reconciliation in Python},
+  year         = {2025},
+  publisher    = {GitHub},
+  howpublished = {\url{https://github.com/dirmeier/reconcile}},
+  note         = {Accessed: 2025-07-15}
+}
 
 @article{ZHANG2024143,
   title = {Discrete forecast reconciliation},
