### 2025-07-16 #8336 @Anubhabsupsi
Thank you very much @robmoss and @earthai-tech for reviewing our submission and for your valuable comments on the software.

Hi @goldingn thanks for the update. I completed the post-review tasks.

The new release (tag **v0.3.3**) is available in **Zenodo** :-
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15974287.svg)](https://doi.org/10.5281/zenodo.15974287)

Let me know if anything else is required!
