### 2025-07-27 #8336 @goldingn
Thanks @Anubhabsupsi! Sorry it took me a little while to get back to this.
Everything seems to be in order with the paper and archive and the article reads very well. I also really like the hand-drawn figures.
