### 2025-07-30 f8f878c
Minor fixes for the paper
