+++ paper/paper.bib
@@ -54,7 +54,7 @@
 }
 
 @article{ZAMBON20241438,
-  title = {Properties of the reconciled distributions for Gaussian and count forecasts},
+  title = {Properties of the reconciled distributions for {G}aussian and count forecasts},
   journal = {International Journal of Forecasting},
   volume = {40},
   number = {4},
@@ -93,7 +93,7 @@
 }
 
 @article{olivares2022hierarchicalforecast,
-  title = {HierarchicalForecast: A Reference Framework for Hierarchical Forecasting in Python},
+  title = {HierarchicalForecast: A Reference Framework for Hierarchical Forecasting in {P}ython},
   author = {Olivares, Kin G. and Garza, Federico and Luo, David and Challú, Cristian and Mergenthaler, Max and Ben Taieb, Souhaib and Wickramasuriya, Shanika L. and Dubrawski, Artur},
   journal = {arXiv preprint arXiv:2207.03517},
   year = {2022},
@@ -102,7 +102,7 @@
 }
 
 @article{gluonTS_Alexandrov_etal2020,
-  title = {GluonTS: Probabilistic and Neural Time Series Modeling in Python},
+  title = {GluonTS: Probabilistic and Neural Time Series Modeling in {P}ython},
   author = {Alexandrov, Alexander and Benidis, Konstantinos and Bohlke-Schneider, Michael and Flunkert, Valentin and Gasthaus, Jan and Januschowski, Tim and Maddix, Danielle C. and Rangapuram, Syama and Salinas, David and Schulz, Jasper and Stella, Lorenzo and Türkmen, Ali Caner and Wang, Yuyang},
   journal = {Journal of Machine Learning Research},
   volume = {21},
@@ -139,9 +139,33 @@
   doi = {10.32614/CRAN.package.fable}
 }
 
+@Manual{ProbReco2020,
+  title        = {ProbReco: Score Optimal Probabilistic Forecast Reconciliation},
+  author       = {Anastasios N. Panagiotelis},
+  year         = {2020},
+  note         = {R package version 0.1.0},
+  url          = {https://github.com/anastasiospanagiotelis/ProbReco},
+}
+
+@misc{discreterecon,
+  author = {AngelPone},
+  title = {{DiscreteRecon}: Discrete Forecast Reconciliation R package},
+  howpublished = {GitHub repository},
+  year = {2023},
+  note = {\url{https://github.com/AngelPone/DiscreteRecon}},
+}
+
+@misc{pyhts2022,
+  title        = {{pyhts}: A Python Package for Hierarchical Forecast Reconciliation},
+  author       = {Bohan Zhang},
+  year         = {2022},
+  howpublished = {\url{https://pypi.org/project/pyhts/}},
+  note         = {Version 0.2.0, BSD‑3‑Clause License, released 11 October 2022}
+}
+
 @misc{scikit-hts,
   author       = {C. Jason Ross},
-  title        = {scikit-hts: Hierarchical Time Series Forecasting with Python},
+  title        = {scikit-hts: Hierarchical Time Series Forecasting with {P}ython},
   year         = {2019},
   publisher    = {GitHub},
   journal      = {GitHub repository},
@@ -150,7 +174,7 @@
 
 @misc{dirmeier2025reconcile,
   author       = {Simon Dirmeier},
-  title        = {reconcile: Hierarchical Time Series Reconciliation in Python},
+  title        = {reconcile: Hierarchical Time Series Reconciliation in {P}ython},
   year         = {2025},
   publisher    = {GitHub},
   howpublished = {\url{https://github.com/dirmeier/reconcile}},
@@ -224,7 +248,7 @@
 }
 
[…]
