### 2025-07-30 7a1944f
Minor fixes for the paper
