+++ paper/paper.bib
@@ -156,7 +156,7 @@
 }
 
 @misc{pyhts2022,
-  title        = {{pyhts}: A Python Package for Hierarchical Forecast Reconciliation},
+  title        = {{pyhts}: A {P}ython Package for Hierarchical Forecast Reconciliation},
   author       = {Bohan Zhang},
   year         = {2022},
   howpublished = {\url{https://pypi.org/project/pyhts/}},
