### 2025-07-30 03f3d77
Update paper.md
